separate headset status dialog into another file
