+++ zen/ui/zigzag-layout.c
@@ -8,7 +8,7 @@
 #include "zen/screen.h"
 #include "zen/server.h"
 #include "zen/ui/nodes/menu-bar.h"
-#include "zen/ui/nodes/vr-modal.h"
+#include "zen/ui/nodes/vr-modal/vr-modal.h"
 
 static void
 zn_zigzag_layout_handle_display_system_changed(
